add code to app, main, login, notfound
