+++ frontend/src/reducers/thoughts.js
